+++ diplomatic/src/rocket/Core.scala
@@ -1,12 +1,11 @@
 // See LICENSE.SiFive for license details.
 
-package freechips.rocketchip.tile
+package org.chipsalliance.rockettile
 
 import Chisel._
-
 import org.chipsalliance.cde.config._
-import freechips.rocketchip.rocket._
 import freechips.rocketchip.util._
+import org.chipsalliance.rocket._
 
 case object XLen extends Field[Int]
 case object MaxHartIdBits extends Field[Int]
