+++ diplomatic/src/rocket/FPU.scala
@@ -1,7 +1,7 @@
 // See LICENSE.Berkeley for license details.
 // See LICENSE.SiFive for license details.
 
-package freechips.rocketchip.tile
+package org.chipsalliance.rockettile
 
 import chisel3._
 import chisel3.util._
@@ -9,8 +9,8 @@ import freechips.rocketchip.util.CompileOptions.NotStrictInferReset
 import chisel3.{DontCare, WireInit, withClock, withReset}
 import chisel3.internal.sourceinfo.SourceInfo
 import org.chipsalliance.cde.config.Parameters
-import freechips.rocketchip.rocket._
-import freechips.rocketchip.rocket.Instructions._
+import org.chipsalliance.rocket._
+import org.chipsalliance.rocket.Instructions._
 import freechips.rocketchip.util._
 import freechips.rocketchip.util.property
