+++ diplomatic/src/rocket/SCIE.scala
@@ -1,6 +1,6 @@
 // See LICENSE.SiFive for license details.
 
-package freechips.rocketchip.scie
+package org.chipsalliance.rocket
 
 import chisel3._
 import chisel3.util.{BitPat, HasBlackBoxInline}
