+++ diplomatic/src/rocket/package.scala
@@ -1,5 +1,5 @@
 // See LICENSE.Berkeley for license details.
 
-package freechips.rocketchip
+package org.chipsalliance
 
 package object rocket extends rocket.constants.ScalarOpConstants with rocket.constants.MemoryOpConstants
